+++ src/components/ErrorNotification.tsx
@@ -1,0 +1,33 @@
+import classNames from 'classnames'
+import React from 'react'
+
+type Props = {
+  errorMessage: string,
+  setErrorMessage: (message: string) => void
+}
+const ErrorNotification: React.FC<Props> = ({
+  errorMessage,
+  setErrorMessage
+}) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={classNames(
+        'notification is-danger is-light has-text-weight-normal',
+        {
+          hidden: !errorMessage,
+        },
+      )}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={() => setErrorMessage('')}
+      />
+      {errorMessage}
+    </div>
+  )
+}
+
+export default ErrorNotification
